+++ scripts/ac5e-helpers.mjs
@@ -3001,11 +3001,11 @@ export function _generateAC5eFlags() {
 		moduleFlags.add(`${moduleFlagScope}.aura.${statusFlagKey}`);
 	}
 
-	const allModesActionTypes = ['all', 'attack', 'check', 'concentration', 'damage', 'death', 'initiative', 'save', 'skill', 'tool'];
-	const noDamageNoInitiativeActionTypes = ['all', 'attack', 'check', 'concentration', 'death', 'save', 'skill', 'tool'];
+	const allModesActionTypes = ['all', 'd20', 'attack', 'check', 'concentration', 'damage', 'death', 'initiative', 'save', 'skill', 'tool'];
+	const noDamageNoInitiativeActionTypes = ['all', 'd20', 'attack', 'check', 'concentration', 'death', 'save', 'skill', 'tool'];
 	const failActionTypes = [...noDamageNoInitiativeActionTypes, 'use'];
-	const noDamageActionTypes = ['all', 'attack', 'check', 'concentration', 'death', 'initiative', 'save', 'skill', 'tool'];
-	const modifierActionTypes = ['attack', 'check', 'concentration', 'damage', 'death', 'initiative', 'save', 'skill', 'tool'];
+	const noDamageActionTypes = ['all', 'd20', 'attack', 'check', 'concentration', 'death', 'initiative', 'save', 'skill', 'tool'];
+	const modifierActionTypes = ['d20', 'attack', 'check', 'concentration', 'damage', 'death', 'initiative', 'save', 'skill', 'tool'];
 	const modifyDCActionTypes = ['save', 'concentration', 'death', 'check', 'skill', 'tool'];
 	const noCriticalActionTypes = ['all', 'attack', 'damage'];
 	const actionTypesByMode = {